+++ java/semantickernel-api/src/main/java/com/microsoft/semantickernel/exceptions/SkillsNotFoundException.java
@@ -1,4 +1,6 @@
 // Copyright (c) Microsoft. All rights reserved.
 package com.microsoft.semantickernel.exceptions;
 
-public class SkillsNotFoundException extends RuntimeException {}
+import com.microsoft.semantickernel.SKException;
+
+public class SkillsNotFoundException extends SKException {}
