+++ java/semantickernel-core/src/main/java/com/microsoft/semantickernel/templateengine/TemplateException.java
@@ -1,6 +1,6 @@
 // Copyright (c) Microsoft. All rights reserved.
 package com.microsoft.semantickernel.templateengine;
 
-import com.microsoft.semantickernel.diagnostics.SKException;
+import com.microsoft.semantickernel.SKException;
 
 public class TemplateException extends SKException {}
